==> 01_Libraries.R <==
mycolor_pallette <- c("steelblue3", "coral3", "thistle4", "aquamarine4","gray1")

library(betapart)
library(grid)
library(mgcv)
library(gratia)
library(patchwork)
library(taxize)
library(plyr)
library(vegan)
library(BiodiversityR)
library(ggsci)
library(data.table)
library(codyn)
library(ggplot2)
library(dplyr)
library(viridis)
library(ape)
library(rlist)
library(RColorBrewer)
library(scales)
library(ggrepel)
library(tidyverse)
library(rlist)
library(harrietr)
library(asbio)
library(qgraph)
library(sf)
library(rgdal)
library(pals)
library(raster)
library(flexclust)
require(data.table)
require(vegan)
require(ggsci)
require(plyr)
require(ggplot2)
library(ncdf4)
library(PCICt)
library(ncdf4.helpers)
library(lubridate)
library(stringr)
library(RColorBrewer)
library(Hmisc)
library(tidyverse)
library(viridis)
library(ggrepel)
library(tidyr)
library(tibble)
library(dplyr)
library(dismo)
library(sp)
library(rlist)
library(raster)
library(maptools)
library(sf)
library(corrplot)
library(chron)
library(fields)
library(rgdal)
library(ggplot2)
library(maptools)
library(grid)
library(maps)
library(ggplot2)
